+++ node/resolvers/Queries/Organizations.ts
@@ -138,7 +138,9 @@ const Organizations = {
     const whereArray = getWhereByStatus({ status })
 
     if (search) {
-      whereArray.push(`name="*${search}*"`)
+      whereArray.push(
+        `(name="*${search}*" OR tradeName="*${search}*" OR id="*${search}*")`
+      )
     }
 
     const where = whereArray.join(' AND ')
@@ -383,7 +385,9 @@ const Organizations = {
       if (search.match(/[a-z\d]+@[a-z]+\.[a-z]{2,3}/gm)) {
         whereArray.push(`b2bCustomerAdmin.email=${search}`)
       } else {
-        whereArray.push(`name="*${search}*"`)
+        whereArray.push(
+          `(name="*${search}*" OR tradeName="*${search}*" OR id="*${search}*")`
+        )
       }
     }